First edition

+++ app/src/main/java/com/example/syriamarket/pojo/products/DataXP.kt
@@ -3,9 +3,9 @@ package com.example.syriamarket.pojo.products
 data class DataXP(
     val _id: String,
     val category: Category,
-    val priceA: Int,
-    val priceB: Int,
-    val priceC: Int,
+    val priceA: Double,
+    val priceB: Double,
+    val priceC: Double,
     val title: String,
     val url: String
 )
